Applied base/ part of patch from Stephen Torri <storri@torri.org>.
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@6494 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ base/AudioLevel.C
@@ -202,7 +202,7 @@ AudioLevel::multiplier_to_fader(float multiplier, int maxLevel, FaderType type)
 {
     if (multiplier == 0.0) return 0;
     float dB = multiplier_to_dB(multiplier);
-    float fader = dB_to_fader(dB, maxLevel, type);
+    int fader = dB_to_fader(dB, maxLevel, type);
     return fader;
 }
